gui: Complete refactor of download functionality
Turns out electron-dl cannot handle multiple simultaneous downloads.
This lead to a rewrite of the entire functionality provided by
electon-dl.

In a way the rewrite is much more beneficial. Now there isn't a lot of
IPC. Node.js standard modules are used for file download.

Currently resume and pause of download is not supported but they are
easy to add in the future.

The gui itself has not been changed much.